add support for ovs when controlled by NSX.
When ovs is managed by NSX there is an integration bridge that will be created for any vif created in that will be wired to a logical switch. This commit adds support for it. Given the existence of an br-int bridge wire the GUEST_IFNAME to the br-int and move it to the namespace. No need to create a veth as that leads to unnecessary copying of bits in the kernel.